move init logger to cli.rs
